refactor: steroids/dev#987 replace absolute @steroidsjs/core import with relative
