+++ src/components/JwtHttpComponent.ts
@@ -1,6 +1,7 @@
-import {logout} from '@steroidsjs/core/actions/auth';
-import BaseHttpComponent from '@steroidsjs/core/components/HttpComponent';
 import axios from 'axios';
+
+import BaseHttpComponent from './HttpComponent';
+import {logout} from '../actions/auth';
 
 export interface IRefreshTokenRequest {
     url: string,
