+++ src/ui/form/InputField/InputField.tsx
@@ -3,12 +3,12 @@
 import {MaskitoOptions} from '@maskito/core';
 import {maskitoDateOptionsGenerator} from '@maskito/kit';
 import {useMaskito} from '@maskito/react';
-import {ISaveCursorPositionDebounceConfig} from '@steroidsjs/core/hooks/useSaveCursorPosition';
 import {InputHTMLAttributes, ReactElement, ReactNode, useCallback, useEffect, useMemo, FocusEvent, MouseEvent} from 'react';
 
 import {INPUT_TYPES_SUPPORTED_SELECTION, useInputFieldWarningByType} from './hooks/useInputFieldWarningByType';
 import {FieldEnum} from '../../../enums';
 import {useComponents, useSaveCursorPosition} from '../../../hooks';
+import {ISaveCursorPositionDebounceConfig} from '../../../hooks/useSaveCursorPosition';
 import fieldWrapper, {IFieldWrapperInputProps, IFieldWrapperOutputProps} from '../Field/fieldWrapper';
 
 export const MASK_PRESETS = {
